remove GLShared::Rect in favor of ivec4

+++ drawing/scene.hpp
@@ -16,7 +16,7 @@ public:
 	virtual void Insert(SMesh obj) = 0;
 	virtual void Erase(SMesh obj) = 0;
 	virtual void SetViewProjectionMat4(const glm::mat4& mat) = 0;
-	virtual void SetViewport(const glm::vec4& rect) = 0;
+	virtual void SetViewport(const glm::ivec4& rect) = 0;
 	virtual void SetNext(SScene next) = 0;
 };
